Login and signup are working. Templates are super ugly
